@@ -5,77 +5,71 @@
   "private": true,
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "watch": "webpack --config webpack.config.js --watch",
-    "start": "webpack --config webpack.config.js --open",
+    "watch": "webpack --watch --config webpack.config.js",
+    "start": "webpack --config webpack.config.js",
     "start1": "webpack-dev-server --config webpack.config.js --open",
     "prod": "webpack --config webpack.config.prod.js",
-    "build": "npm run test1 test2 test && webpack-dev-server --config webpack.config.js --open && echo \"Npm started\"",
+    "buil": "npm run test1 test2 test && webpack-dev-server --config webpack.config.js --open && echo \"Npm started\"",
     "test1": "npm i --save-dev webpack webpack-cli",
-    "test2": "npm i"
+    "test2": "npm i",
+    "linter": "npm run fork-ts-checker-webpack-plugin",
+    "build": "webpack"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Chugou9/expenses.git"
   },
   "keywords": [],
-  "author": "",
+  "author": "chugou",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/Chugou9/expenses/issues"
   },
   "homepage": "https://github.com/Chugou9/expenses#readme",
   "devDependencies": {
-    "@babel/core": "^7.7.2",
-    "@babel/plugin-proposal-class-properties": "^7.7.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
-    "@babel/preset-env": "^7.7.1",
-    "@babel/preset-react": "^7.7.0",
-    "@babel/preset-typescript": "^7.7.2",
-    "@types/cometd": "^4.0.6",
-    "@types/jest": "^24.0.23",
+    "@emotion/babel-plugin": "^11.0.0",
+    "@types/jest": "^27.0.3",
     "@types/lodash.clonedeep": "^4.5.6",
     "@types/lodash.isempty": "^4.4.6",
-    "@types/node": "^12.12.7",
-    "@types/react": "^16.9.11",
-    "@types/react-dom": "^16.9.4",
-    "@types/react-router-dom": "^5.1.2",
-    "babel-cli": "^6.18.0",
-    "babel-loader": "^8.0.6",
-    "babel-plugin-emotion": "^10.0.27",
-    "babel-preset-react": "^6.24.1",
-    "clean-webpack-plugin": "^3.0.0",
-    "css-loader": "^3.2.0",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^5.0.2",
-    "fork-ts-checker-webpack-plugin": "^3.1.0",
-    "html-webpack-plugin": "^3.2.0",
-    "less": "^3.10.3",
-    "less-loader": "^5.0.0",
+    "@types/node": "^17.0.1",
+    "@types/react": "^17.0.37",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.2",
+    "@typescript-eslint/parser": "^5.8.1",
+    "clean-webpack-plugin": "^4.0.0",
+    "css-loader": "^6.5.1",
+    "eslint": "^8.6.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-webpack-plugin": "^3.1.1",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.5.0",
+    "less": "^4.1.2",
+    "less-loader": "^10.2.0",
     "lodash.isempty": "^4.4.0",
-    "source-map-loader": "^0.2.4",
-    "style-loader": "^1.0.0",
-    "terser-webpack-plugin": "^2.2.1",
-    "ts-loader": "^6.2.1",
-    "tslint": "^5.20.1",
-    "typescript": "^3.7.2",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.9.0"
+    "source-map-loader": "^3.0.0",
+    "style-loader": "^3.3.1",
+    "terser-webpack-plugin": "^5.3.0",
+    "ts-loader": "^9.2.6",
+    "typescript": "^4.5.4",
+    "webpack": "^5.65.0",
+    "webpack-bundle-analyzer": "^4.5.0",
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-server": "^4.6.0"
   },
   "dependencies": {
-    "@emotion/core": "^10.0.27",
-    "@fortawesome/fontawesome-svg-core": "^1.2.25",
-    "@fortawesome/free-solid-svg-icons": "^5.11.2",
-    "@fortawesome/react-fontawesome": "^0.1.7",
+    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@fortawesome/react-fontawesome": "^0.1.16",
+    "axios": "^0.24.0",
     "bootstrap": "^3.4.1",
-    "cometd": "3.1.5",
-    "emotion": "^10.0.27",
+    "fork-ts-checker-notifier-webpack-plugin": "^4.0.0",
+    "fork-ts-checker-webpack-plugin": "^6.5.0",
     "lodash.clonedeep": "^4.5.0",
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0",
-    "react-router-dom": "^5.1.2",
-    "react-spinners": "^0.8.0",
-    "superagent": "^5.1.3",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.1",
+    "react-spinners": "^0.11.0",
+    "superagent": "^6.1.0",
     "superagent-no-cache": "^0.1.1"
   }
 }
