@@ -1,6 +1,6 @@
 import {AuthContext} from "Common/Contexts/Auth.context";
 import * as React from "react";
-import {Route, Switch} from 'react-router-dom';
+import {Route, Routes} from 'react-router-dom';
 import {CONFIGURATION} from "../Config/Config";
 import {ROUTES} from "../Consts";
 import AuthorisationPage from 'Modules/Authorisation/AuthorisationPage';
@@ -9,28 +9,26 @@ import AuthorisationPage from 'Modules/Authorisation/AuthorisationPage';
  * Свитч, который раскидывает юзера по конкретным компонентам.
  */
 function RouterSwitch() {
-    const {token} = React.useContext(AuthContext);
+    const {userId} = React.useContext(AuthContext);
 
     return (
-        <Switch>
-            {!token && (
+        <Routes>
+            {!userId && (
                 <Route
                     key="auth"
                     path={ROUTES.AUTHORISATION.PATH}
-                    exact
-                    children={<AuthorisationPage />}
+                    element={<AuthorisationPage />}
                 />
             )}
 
-            {!!token && CONFIGURATION.map((route, index) => (
+            {!!userId && CONFIGURATION.map((route, index) => (
                 <Route
                     key={index}
-                    exact={route.exact}
                     path={route.path}
-                    children={route.main}
+                    element={route.main}
                 />
             ))}
-        </Switch>
+        </Routes>
     );
 }
 
