@@ -1,5 +1,4 @@
 import * as React from "react";
-import { isFunction } from "util";
 
 /**
  * Собственные свойства компонента ввода.
@@ -33,8 +32,8 @@ export function Input({
      */
     const handleChange = (e: React.SyntheticEvent<HTMLInputElement>) => {
         let value = e.currentTarget.value;
-        
-        isFunction(onChange) && onChange(value ? value : null);
+
+        onChange && onChange(value ? value : null);
     };
 
     return (
