@@ -3,9 +3,8 @@ import {createContext} from 'react';
 function noop() {}
 
 interface IAuthContext {
-  token: string | null,
   userId: string | null,
-  login: (jwt: string | null, id: string | null) => void,
+  login: (id: string | null) => void,
   logout: () => void,
   isAuthentificated: boolean
 }
@@ -14,7 +13,6 @@ interface IAuthContext {
  * Контекст для прокидывания данных авторизации.
  */
 export const AuthContext = createContext<IAuthContext>({
-  token: null,
   userId: null,
   login: noop,
   logout: noop,
