@@ -1,23 +1,30 @@
 import { SERVER_PATH } from "Consts/REST";
-import { POST } from "Utils/REST";
+import { GET, POST } from "Utils/REST";
 
 export class AuthServices {
     constructor() {}
 
-    logIn(request: any, setLogin: (jwt: string, userId: string) => void) {
-        return POST(`${SERVER_PATH}/api/user/login`, request).then(
-            (response: {token: string, userId: string}) => {
+    logIn(request: any, setLogin: (userId: string) => void) {
+        return POST<{userId: string}>(`${SERVER_PATH}/api/user/login`, request).then(
+            (response) => {
                 console.log('json', response);
-                setLogin(response.token, response.userId)
+                setLogin(response.userId)
             },
             (err) => console.log('Login error', err)
         );
     }
 
-    register(request: any, setLogin: (jwt: string, userId: string) => void) {
-        return POST(`${SERVER_PATH}/api/user/register`, request).then(
-            (response) => setLogin(response.jwt, response.userId),
+    register(request: any, setLogin: (userId: string) => void) {
+        return POST<{userId: string}>(`${SERVER_PATH}/api/user/register`, request).then(
+            (response) => setLogin(response.userId),
             (err) => console.log('Login error', err)
+        );
+    }
+
+    test() {
+        return GET(`${SERVER_PATH}/api`).then(
+            response => console.log('response', response),
+            err => console.log('error', err)
         );
     }
 }
