@@ -1,4 +1,4 @@
 /**
  * Путь до сервера.
  */
-export const SERVER_PATH = 'http://localhost';
+export const SERVER_PATH = location.origin;
